+++ CHANGELOG.md
@@ -2951,3 +2951,4 @@
 - [2026-07-20 13:43] feat(cache): add bi-directional event bus handler for cache
 - [2026-07-20 14:14] docs(storage): update readme and API contracts for storage
 - [2026-07-20 14:24] fix(security): fix memory leak by cancelling security subscription
+- [2026-07-20 15:08] perf(security): memoize expensive computations in security